fix: misc

+++ api/models/DailyChallenge.js
@@ -24,11 +24,11 @@ export const DailyChallenge = sequelize.define('DailyChallenge', {
         allowNull: false,
     },
     time: {
-        type: DataTypes.DOUBLE,
+        type: DataTypes.INTEGER,
         allowNull: false,
     },
     points: {
-        type: DataTypes.DOUBLE,
+        type: DataTypes.INTEGER,
         allowNull: false,
     },
     is_solved: {
